requirements.txt
streamlit==1.29.0
pandas==1.3.3
scikit-learn==0.24.2
pillow==8.4.0
matplotlib==3.4.3
plotly==5.3.1
seaborn==0.11.2
